make x, y, a1, a2 double

+++ figure.h
@@ -4,11 +4,11 @@
 #include "data/list.h"
 
 typedef struct {
-	int x;
-	int y;
+	double x;
+	double y;
 
-	int a1;
-	int a2;
+	double a1;
+	double a2;
 
 	int type;
